Revert branch to state at 7a1e512
