+++ front-end/src/components/Events.tsx
@@ -14,8 +14,7 @@ import { useSlideScroll } from './core/MultiView'
 import Row from './core/Row'
 import RowSelect from './core/RowSelect'
 import Spacer from './core/Spacer'
-import CompactEvents from './events/CompactEvents'
-import Event from './events/Event'
+import Event, { EventInfo } from './events/Event'
 import EventEditor from './events/EventEditor'
 
 type EventsFilter = 'All' | 'Bookmarked' | 'Mine'
@@ -178,4 +177,40 @@ function Events({ events, editEvent }: {events: readonly DayjsEvent[], editEvent
             )
         })}
     </>
+}
+
+function CompactEvents({ events, editEvent }: { events: readonly DayjsEvent[], editEvent: (eventId: string) => void }) {
+    const { hashState, setHashState, getHashStateString } = useHashState()
+
+    const viewingEvent = useMemo(() => events.find(e => e.event_id === hashState?.viewingEventDetails), [events, hashState?.viewingEventDetails])
+
+    return (
+        <>
+            <div className='compactEvents'>
+                <a className='headings'>
+                    <div className='time'>When</div>
+                    <div className='name'>What</div>
+                </a>
+                {events.map(e => {
+                    return (
+                        <a
+                            href={'#' + getHashStateString({ viewingEventDetails: e.event_id })}
+                            key={e.event_id}
+                        >
+                            <div className='time'>{e.start_datetime.format('ddd h:mma')}</div>
+                            <div className='name'>{e.name}</div>
+                        </a>
+                    )
+                })}
+            </div>
+
+            <Modal isOpen={viewingEvent != null} onClose={() => setHashState({ viewingEventDetails: undefined })} side='right'>
+                {() =>
+                    viewingEvent &&
+                        <div style={{ padding: 20 }}>
+                            <EventInfo event={viewingEvent} editEvent={editEvent} />
+                        </div>}
+            </Modal>
+        </>
+    )
 }
