+++ front-end/src/components/events/EventOverlapModal.tsx
@@ -1,12 +1,10 @@
 import React from 'react'
 
 import { DayjsEvent } from '../../hooks/useStore'
-import { doNothing } from '../../utils'
 import Button from '../core/Button'
 import Col from '../core/Col'
 import Modal from '../core/Modal'
 import Spacer from '../core/Spacer'
-import CompactEvents from './CompactEvents'
 
 type Props = {
     isOpen: boolean,
@@ -21,15 +19,20 @@ export default React.memo(({ isOpen, onClose, onConfirm, overlappingEvents }: Pr
             {() => (
                 <Col align='center' justify='center' padding={20} pageLevel>
                     <div style={{ fontSize: 22, textAlign: 'center' }}>
-                        Your event overlaps with the following. Schedule it anyway?
+                        This event overlaps with:
+                        {overlappingEvents.map(event => (
+                            <div key={event.event_id} style={{ fontSize: 16, marginTop: 8 }}>
+                                &quot;{event.name}&quot;
+                                <br />
+                                {event.start_datetime.format('MMM D, h:mm A')} -{' '}
+                                {event.end_datetime?.format('h:mm A') ?? 'onwards'}
+                                <br />
+                                {event.event_site_location_name}
+                            </div>
+                        ))}
+                        <br />
+                        Want to schedule it anyway?
                     </div>
-
-                    <Spacer size={16} />
-
-                    <CompactEvents
-                        events={overlappingEvents}
-                        editEvent={doNothing}
-                    />
 
                     <Spacer size={16} />
